+++ attributes/default.rb
@@ -48,6 +48,9 @@ default['openstack']['image']['notification_driver'] = 'noop'
 # Set the number of api workers
 default['openstack']['image']['api']['workers'] = [8, node['cpu']['total'].to_i].min
 
+# Set the number of registry workers
+default['openstack']['image']['registry']['workers'] = [8, node['cpu']['total'].to_i].min
+
 # Return the URL that references where the data is stored on the backend.
 default['openstack']['image']['api']['show_image_direct_url'] = 'False'