+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This file is used to list changes made in each version of cookbook-openstack-ima
 * Enable authtoken configurations - including cafile, insecure, memcached_servers, memcache_security_strategy, memcache_secret_key and hash_algorithms
 * Fix metadata version constraint for common
 * Update conf files with scerect information from mode 644 to 640
+* Add attribute for glance-registry workers
 
 ## 9.2.0
 * python_packages database client attributes have been migrated to the -common cookbook